LPD-75170 refactor: move all left files from ai-hub-impl to ai-hub-nexus-impl
