+++ modules/dxp/apps/ai-hub-nexus/ai-hub-nexus-impl/src/main/java/com/liferay/ai/hub/nexus/internal/memory/ChatMemoryProviderUtil.java
@@ -1,0 +1,31 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+package com.liferay.ai.hub.nexus.internal.memory;
+
+import dev.langchain4j.memory.chat.MessageWindowChatMemory;
+import dev.langchain4j.store.memory.chat.InMemoryChatMemoryStore;
+
+/**
+ * @author Feliphe Marinho
+ * @author João Victor Alves
+ */
+public class ChatMemoryProviderUtil {
+
+	public static MessageWindowChatMemory provide(Object memoryId) {
+		return MessageWindowChatMemory.builder(
+		).chatMemoryStore(
+			_inMemoryChatMemoryStore
+		).id(
+			memoryId
+		).maxMessages(
+			30
+		).build();
+	}
+
+	private static final InMemoryChatMemoryStore _inMemoryChatMemoryStore =
+		new InMemoryChatMemoryStore();
+
+}
