--- modules/dxp/apps/ai-hub/ai-hub-impl/bnd.bnd
@@ -1,18 +1,0 @@
-Bundle-Name: Liferay AI Hub Implementation
-Bundle-SymbolicName: com.liferay.ai.hub.impl
-Bundle-Version: 1.0.6
-Import-Package:\
-	!com.fasterxml.jackson.datatype.*,\
-	\
-	!org.bouncycastle.jsse.*,\
-	\
-	!org.conscrypt.*,\
-	\
-	!org.openjsse.javax.net.ssl.*,\
-	!org.openjsse.net.ssl.*,\
-	\
-	android.*;resolution:=optional,\
-	\
-	*
--dsannotations-options: inherit
--fixupmessages: Classes found in the wrong directory: ...;is:=ignore
